requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-telegram-bot==21.6
python-dotenv==1.0.0
gunicorn==21.2.0